Updated Home

+++ ParentsBank/Controllers/WishlistItemsController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
@@ -15,20 +16,21 @@ namespace ParentsBank.Controllers
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: WishlistItems
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            var wishlistItems = db.WishlistItems.Include(w => w.Account);
-            return View(wishlistItems.ToList());
+            string user = User.Identity.Name;
+            var wishlistItems = db.WishlistItems.Include(w => w.Account).Where(w => w.Account.Owner.ToLower() == user.ToLower()).Union(db.WishlistItems.Include(w => w.Account).Where(w => w.Account.Recipient.ToLower() == user.ToLower()));
+            return View(await wishlistItems.ToListAsync());
         }
 
         // GET: WishlistItems/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            WishlistItem wishlistItem = db.WishlistItems.Find(id);
+            WishlistItem wishlistItem = await db.WishlistItems.FindAsync(id);
             if (wishlistItem == null)
             {
                 return HttpNotFound();
@@ -39,7 +41,9 @@ namespace ParentsBank.Controllers
         // GET: WishlistItems/Create
         public ActionResult Create()
         {
-            ViewBag.AccountId = new SelectList(db.Accounts, "Id", "Owner");
+            string user = User.Identity.Name;
+            List<AccountDetails> list = db.Accounts.Where(model => model.Owner.ToLower() == user.ToLower()).Union(db.Accounts.Where(model => model.Recipient.ToLower() == user.ToLower())).ToList();
+            ViewBag.AccountId = PopulateAccountSelectItems(list, null);
             return View();
         }
 
@@ -48,32 +52,37 @@ namespace ParentsBank.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,AccountId,DateAdded,Cost,Description,Link,Purchased")] WishlistItem wishlistItem)
+        public async Task<ActionResult> Create([Bind(Include = "Id,AccountId,DateAdded,Cost,Description,Link,Purchased")] WishlistItem wishlistItem)
         {
+            wishlistItem.DateAdded = DateTime.Now;
             if (ModelState.IsValid)
             {
                 db.WishlistItems.Add(wishlistItem);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
-            ViewBag.AccountId = new SelectList(db.Accounts, "Id", "Owner", wishlistItem.AccountId);
+            string user = User.Identity.Name;
+            List<AccountDetails> list = db.Accounts.Where(model => model.Owner.ToLower() == user.ToLower()).Union(db.Accounts.Where(model => model.Recipient.ToLower() == user.ToLower())).ToList();
+            ViewBag.AccountId = PopulateAccountSelectItems(list, wishlistItem.AccountId);
             return View(wishlistItem);
         }
 
         // GET: WishlistItems/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            WishlistItem wishlistItem = db.WishlistItems.Find(id);
+            WishlistItem wishlistItem = await db.WishlistItems.FindAsync(id);
             if (wishlistItem == null)
             {
                 return HttpNotFound();
             }
-            ViewBag.AccountId = new SelectList(db.Accounts, "Id", "Owner", wishlistItem.AccountId);
+            string user = User.Identity.Name;
+            List<AccountDetails> list = db.Accounts.Where(acct => acct.Id == wishlistItem.AccountId).ToList();
+            ViewBag.AccountId = PopulateAccountSelectItems(list, null);
             return View(wishlistItem);
         }
 
@@ -82,26 +91,28 @@ namespace ParentsBank.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,AccountId,DateAdded,Cost,Description,Link,Purchased")] WishlistItem wishlistItem)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,AccountId,DateAdded,Cost,Description,Link,Purchased")] WishlistItem wishlistItem)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(wishlistItem).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
-            ViewBag.AccountId = new SelectList(db.Accounts, "Id", "Owner", wishlistItem.AccountId);
+            string user = User.Identity.Name;
+            List<AccountDetails> list = db.Accounts.Where(acct => acct.Id == wishlistItem.AccountId).ToList();
+            ViewBag.AccountId = PopulateAccountSelectItems(list, null);
             return View(wishlistItem);
         }
 
         // GET: WishlistItems/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            WishlistItem wishlistItem = db.WishlistItems.Find(id);
+            WishlistItem wishlistItem = await db.WishlistItems.FindAsync(id);
             if (wishlistItem == null)
             {
                 return HttpNotFound();
@@ -112,12 +123,20 @@ namespace ParentsBank.Controllers
         // POST: WishlistItems/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            WishlistItem wishlistItem = db.WishlistItems.Find(id);
-            db.WishlistItems.Remove(wishlistItem);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            WishlistItem wishlistItem = await db.WishlistItems.FindAsync(id);
+            string user = User.Identity.Name;
+            if (wishlistItem.Account.Recipient.ToLower().Equals(user.ToLower()))
+            {
+                ModelState.AddModelError("","Recipient cannot delete a wishlist item.");
+                return View(wishlistItem);
+            }
+            else {
+                db.WishlistItems.Remove(wishlistItem);
+                await db.SaveChangesAsync();
+                return RedirectToAction("Index");
+            }
         }
 
         protected override void Dispose(bool disposing)
@@ -127,6 +146,23 @@ namespace ParentsBank.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        public List<SelectListItem> PopulateAccountSelectItems(List<AccountDetails> list, int? accountId)
+        {
+            List<SelectListItem> selectList = new List<SelectListItem>();
+            foreach (AccountDetails account in list)
+            {
+                SelectListItem item = new SelectListItem();
+                item.Text = account.Owner.ToLower() + ", " + account.Recipient.ToLower();
+                item.Value = account.Id.ToString();
+                if (accountId == account.Id)
+                {
+                    item.Selected = true;
+                }
+                selectList.Add(item);
+            }
+            return selectList;
         }
     }
 }
